feat: redo layout, improve list view
