+++ frontend/components.d.ts
@@ -41,6 +41,7 @@ declare module 'vue' {
     ColorPopover: typeof import('./src/components/ColorPopover.vue')['default']
     ColourPicker: typeof import('./src/components/EspressoIcons/Colour-picker.vue')['default']
     Comment: typeof import('./src/components/EspressoIcons/Comment.vue')['default']
+    ContextMenu: typeof import('./src/components/ContextMenu.vue')['default']
     CTADeleteDialog: typeof import('./src/components/CTADeleteDialog.vue')['default']
     CustomListRow: typeof import('./src/components/CustomListRow.vue')['default']
     Delete: typeof import('./src/components/DocEditor/icons/delete.vue')['default']
@@ -106,6 +107,7 @@ declare module 'vue' {
     LucideAxe: typeof import('~icons/lucide/axe')['default']
     LucideCheck: typeof import('~icons/lucide/check')['default']
     LucideEye: typeof import('~icons/lucide/eye')['default']
+    LucideEyeOff: typeof import('~icons/lucide/eye-off')['default']
     LucideHome: typeof import('~icons/lucide/home')['default']
     LucideLogOut: typeof import('~icons/lucide/log-out')['default']
     LucideSearch: typeof import('~icons/lucide/search')['default']
