+++ frontend/src/App.vue
@@ -1,96 +1,42 @@
 <template>
-  <template v-if="$route.meta.errorPage">
-    <router-view :key="$route.fullPath" v-slot="{ Component }">
-      <component :is="Component" id="currentPage" ref="currentPage" />
-    </router-view>
-  </template>
-  <template v-else>
-    <!-- Main container no scroll -->
-    <div class="flex w-screen h-screen antialiased overflow-hidden">
-      <!-- Main container with scroll -->
+  <div class="w-screen h-screen antialiased">
+    <div class="bg-black text-white text-sm text-center py-2 sm:hidden">
+      Drive works best on desktop.
+    </div>
+    <div v-if="isLoggedIn || $route.meta.allowGuest" class="flex">
+      <Sidebar
+        v-if="isLoggedIn && !['Teams', 'Setup'].includes($route.name)"
+        class="hidden sm:block"
+      />
       <div
-        v-if="
-          !isLoggedIn && ['File', 'Folder', 'Document'].includes($route.name)
-        "
-        class="absolute flex gap-2 bottom-4 right-4 bg-white opacity-90 shadow-md text-gray-800 px-2 py-2 rounded-sm text-sm"
+        id="dropzone"
+        class="flex flex-col h-screen flex-grow overflow-hidden"
       >
-        <FrappeDriveLogo class="w-8 h-8 rounded my-auto" />
-        <div>
-          Powered by
-          <span class="font-bold my-auto">Frappe Drive</span>
-          <br />
-          <Button
-            variant="outline"
-            size="sm"
-            class="mt-2"
-            @click="$router.push({ name: 'Signup' })"
-          >
-            Join now
-          </Button>
-        </div>
-      </div>
-      <div class="h-full w-full flex flex-col">
-        <div class="bg-black text-white text-sm text-center py-2 sm:hidden">
-          Drive works best on desktop.
-        </div>
-        <SearchPopup
-          v-if="isLoggedIn && showSearchPopup"
-          v-model="showSearchPopup"
-        />
-        <div
-          v-if="isLoggedIn || $route.meta.allowGuest"
-          class="flex flex-col h-full overflow-hidden sm:flex-row"
-        >
-          <Sidebar
-            v-if="isLoggedIn && !['Teams', 'Setup'].includes($route.name)"
-            class="hidden sm:block"
-          />
-          <div id="dropTarget" class="h-full w-full overflow-hidden">
-            <Navbar v-if="$route.name == 'File' || $route.name == 'Document'" />
-            <div class="flex w-full h-full overflow-hidden">
-              <!-- Find a better way to handle the height overflow here (52px is the Navbar) -->
-              <!-- what on mars is he talking about? -->
-              <div
-                class="flex w-full overflow-hidden"
-                :class="
-                  $route.name == 'File' || $route.name == 'Document'
-                    ? 'h-[calc(100dvh-105px)] sm:h-[calc(100dvh-52px)]'
-                    : 'h-full sm:h-full'
-                "
-              >
-                <router-view :key="$route.fullPath" v-slot="{ Component }">
-                  <component
-                    :is="Component"
-                    id="currentPage"
-                    ref="currentPage"
-                  />
-                </router-view>
-              </div>
-              <InfoSidebar v-if="showInfoSidebar" />
-            </div>
-          </div>
-          <BottomBar
-            v-if="isLoggedIn"
-            class="fixed bottom-0 w-full sm:hidden"
-          />
-        </div>
-        <router-view v-else :key="$route.fullPath" v-slot="{ Component }">
-          <component :is="Component" id="currentPage" ref="currentPage" />
+        <router-view :key="$route.fullPath" v-slot="{ Component }">
+          <component :is="Component" class="flex-grow" />
         </router-view>
       </div>
+
+      <InfoSidebar v-if="showInfoSidebar" />
+
+      <BottomBar v-if="isLoggedIn" class="fixed bottom-0 w-full sm:hidden" />
     </div>
-    <Transition
-      enter-active-class="transition duration-[150ms] ease-[cubic-bezier(.21,1.02,.73,1)]"
-      enter-from-class="translate-y-1 opacity-0"
-      enter-to-class="translate-y-0 opacity-100"
-      leave-active-class="transition duration-[150ms] ease-[cubic-bezier(.21,1.02,.73,1)]"
-      leave-from-class="translate-y-0 opacity-100"
-      leave-to-class="translate-y-1 opacity-0"
-    >
-      <UploadTracker v-if="showUploadTracker" />
-    </Transition>
-    <Toasts />
-  </template>
+    <router-view v-else :key="$route.fullPath" v-slot="{ Component }">
+      <component :is="Component" />
+    </router-view>
+  </div>
+  <SearchPopup v-if="isLoggedIn && showSearchPopup" v-model="showSearchPopup" />
+  <Transition
+    enter-active-class="transition duration-[150ms] ease-[cubic-bezier(.21,1.02,.73,1)]"
+    enter-from-class="translate-y-1 opacity-0"
+    enter-to-class="translate-y-0 opacity-100"
+    leave-active-class="transition duration-[150ms] ease-[cubic-bezier(.21,1.02,.73,1)]"
+    leave-from-class="translate-y-0 opacity-100"
+    leave-to-class="translate-y-1 opacity-0"
+  >
+    <UploadTracker v-if="showUploadTracker" />
+  </Transition>
+  <Toasts />
 </template>
 <script setup>
 import Navbar from "@/components/Navbar.vue"
